test: remove unreachable runtime api helpers
