+++ docs/internal/baselines/deadcode-baseline.txt
@@ -189,21 +189,6 @@ tests/functional/replay_contracts/short_helpers_test.go: unreachable func: asser
 tests/functional/replay_contracts/short_helpers_test.go: unreachable func: factoryRelationsValue
 tests/functional/replay_contracts/short_helpers_test.go: unreachable func: replayEventCount
 tests/functional/replay_contracts/short_helpers_test.go: unreachable func: strPtr
-tests/functional/runtime_api/api_generated_smoke_test.go: unreachable func: functionalServerBase
-tests/functional/runtime_api/api_generated_smoke_test.go: unreachable func: mustSubmitWorkImageItem
-tests/functional/runtime_api/api_generated_smoke_test.go: unreachable func: mustSubmitWorkTextItem
-tests/functional/runtime_api/api_generated_smoke_test.go: unreachable func: requireGeneratedWorkByTrace
-tests/functional/runtime_api/api_generated_smoke_test.go: unreachable func: stageGeneratedSubmitWorkFile
-tests/functional/runtime_api/api_generated_smoke_test.go: unreachable func: waitForGeneratedWorkTypeComplete
-tests/functional/runtime_api/api_runtime_config_alignment_smoke_support_test.go: unreachable func: assertRuntimeConfigAlignmentResourceManifest
-tests/functional/runtime_api/api_runtime_config_alignment_support_test.go: unreachable func: runtimeConfigAlignmentComparableFactory
-tests/functional/runtime_api/api_session_invocation_test.go: unreachable func: textInvocationRequest
-tests/functional/runtime_api/external_support_test.go: unreachable func: simplePipelineConfig
-tests/functional/runtime_api/functional_server_test.go: unreachable func: withWorkerCommands
-tests/functional/runtime_api/helpers_test.go: unreachable func: withClock
-tests/functional/runtime_api/helpers_test.go: unreachable func: withInvocationMetricsRecorder
-tests/functional/runtime_api/helpers_test.go: unreachable func: withSubmissionRecorder
-tests/functional/runtime_api/runtime_support_test.go: unreachable func: requireGeneratedSchemaRunStartedPayload
 tests/functional/sessions/execution/helpers_test.go: unreachable func: pollHostedCLIInvocationObservations
 tests/functional/sessions/execution/helpers_test.go: unreachable func: tryWaitForTerminalWorkPrimaryTextDuringInvocation
 tests/functional/sessions/execution/helpers_test.go: unreachable func: waitForTerminalWorkPrimaryText
