+++ tests/functional/runtime_api/runtime_support_test.go
@@ -2,7 +2,6 @@ package runtime_api
 
 import (
 	"fmt"
-	"testing"
 	"time"
 
 	"github.com/portpowered/infinite-you/pkg/services/work"
@@ -56,24 +55,4 @@ func stringValueFromFunctionalPtr[T ~string](value *T) string {
 		return ""
 	}
 	return string(*value)
-}
-
-func requireGeneratedSchemaRunStartedPayload(t *testing.T, events []factoryapi.FactoryEvent) factoryapi.RunRequestEventPayload {
-	t.Helper()
-
-	for _, event := range events {
-		if event.Type != factoryapi.FactoryEventTypeRunRequest {
-			continue
-		}
-		payload, err := event.Payload.AsRunRequestEventPayload()
-		if err != nil {
-			t.Fatalf("decode run-request payload %q: %v", event.Id, err)
-		}
-		if payload.Factory.WorkTypes == nil || len(*payload.Factory.WorkTypes) == 0 {
-			t.Fatalf("run-request payload factory missing work types: %#v", payload.Factory)
-		}
-		return payload
-	}
-	t.Fatalf("recorded events missing RUN_REQUEST: %#v", functionalEventTypes(events))
-	return factoryapi.RunRequestEventPayload{}
 }
